Refactor CLI into modular producer and consumer directories
Split single cli.rs into specialized modules with shared functionality:
• Producer CLI commands in producer/cli.rs
• Consumer CLI commands in consumer/cli.rs (includes offset operations)
• Combined CLI interface maintaining backward compatibility
• Renamed client implementation files to avoid module naming conflicts
