+++ src/http/producer/mod.rs
@@ -1,0 +1,5 @@
+pub mod cli;
+pub mod client;
+
+pub use cli::*;
+pub use client::*;
